requirements.txt
alembic==1.14.0
APIFlask==3.0.2
celery==5.4.0
flask==3.1.2
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
marshmallow-sqlalchemy==1.4.2
mysqlclient==2.2.7
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.3
